requirements.txt
# Core dependencies
streamlit>=1.31.0
langchain>=0.2.14
langgraph>=0.2.19
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM Providers
ollama>=0.3.1
groq>=0.4.0
google-generativeai>=0.3.0

# Search Tools
duckduckgo-search>=6.2.12
wikipedia>=1.4.0

# Additional dependencies can be installed from:
# - requirements-chroma.txt for Chroma vector store
# - requirements-qdrant.txt for Qdrant vector store
# - requirements-faiss.txt for FAISS vector store